+++ airflow/providers/microsoft/azure/CHANGELOG.rst
@@ -48,6 +48,9 @@ Misc
    * ``Use 'AirflowProviderDeprecationWarning' in providers (#30975)``
    * ``Upgrade ruff to 0.0.262 (#30809)``
    * ``Add full automation for min Airflow version for providers (#30994)``
+   * ``Use '__version__' in providers not 'version' (#31393)``
+   * ``Fixing circular import error in providers caused by airflow version check (#31379)``
+   * ``Prepare docs for May 2023 wave of Providers (#31252)``
 
 6.0.0
 ......